requirements.txt
asn1crypto==0.24.0
boto3==1.16.23
botocore==1.19.23
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
entrypoints==0.3
flake8==3.7.9
google-api-core==1.7.0
google-auth==1.6.2
google-auth-oauthlib==0.2.0
google-cloud-bigquery==1.8.1
google-cloud-core==0.29.1
google-resumable-media==0.3.2
googleapis-common-protos==1.5.5
idna==2.8
jeepney==0.4
jmespath==0.10.0
keyring==17.1.1
mccabe==0.6.1
numpy==1.15.4
oauthlib==2.1.0
pandas==0.23.4
pandas-gbq==0.8.0
protobuf==3.6.1
psycopg2-binary==2.7.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pyodbc==4.0.25
python-dateutil==2.7.5
pytz==2018.7
PyYAML==4.2b4
requests==2.21.0
requests-oauthlib==1.0.0
rsa==4.0
s3transfer==0.3.3
SecretStorage==3.1.0
six==1.12.0
SQLAlchemy==1.2.15
urllib3==1.26.2
